Move useFetchSloList into main hooks folder
